feat: Add support for setting axis unit in matplotlib plots
This commit adds a new function `set_axis_unit` to the `_matplotlib_compat` module. This function allows users to set the scale of the specified axis to a target unit in matplotlib plots. It takes parameters such as the axis index, target unit, matplotlib axis object, and precision for adjusting the axis labels. The function also updates the axis label with the target unit.

The commit also includes a new test case `test_set_axis_unit` in the `_matplotlib_compat_test` module to verify the functionality of the `set_axis_unit` function.